Changed Kernel ABI to match the unboxed primitive shapes for eco dialect.

+++ elm-kernel-cpp/src/core/Bitwise.cpp
@@ -2,7 +2,7 @@
 
 namespace Elm::Kernel::Bitwise {
 
-int32_t and_(int32_t a, int32_t b) {
+int64_t and_(int64_t a, int64_t b) {
     /*
      * JS: var _Bitwise_and = F2(function(a, b) { return a & b; });
      *
@@ -15,7 +15,7 @@ int32_t and_(int32_t a, int32_t b) {
     return a & b;
 }
 
-int32_t or_(int32_t a, int32_t b) {
+int64_t or_(int64_t a, int64_t b) {
     /*
      * JS: var _Bitwise_or = F2(function(a, b) { return a | b; });
      *
@@ -28,7 +28,7 @@ int32_t or_(int32_t a, int32_t b) {
     return a | b;
 }
 
-int32_t xor_(int32_t a, int32_t b) {
+int64_t xor_(int64_t a, int64_t b) {
     /*
      * JS: var _Bitwise_xor = F2(function(a, b) { return a ^ b; });
      *
@@ -41,7 +41,7 @@ int32_t xor_(int32_t a, int32_t b) {
     return a ^ b;
 }
 
-int32_t complement(int32_t a) {
+int64_t complement(int64_t a) {
     /*
      * JS: function _Bitwise_complement(a) { return ~a; }
      *
@@ -54,7 +54,7 @@ int32_t complement(int32_t a) {
     return ~a;
 }
 
-int32_t shiftLeftBy(int32_t offset, int32_t a) {
+int64_t shiftLeftBy(int64_t offset, int64_t a) {
     /*
      * JS: var _Bitwise_shiftLeftBy = F2(function(offset, a) { return a << offset; });
      *
@@ -68,7 +68,7 @@ int32_t shiftLeftBy(int32_t offset, int32_t a) {
     return a << offset;
 }
 
-int32_t shiftRightBy(int32_t offset, int32_t a) {
+int64_t shiftRightBy(int64_t offset, int64_t a) {
     /*
      * JS: var _Bitwise_shiftRightBy = F2(function(offset, a) { return a >> offset; });
      *
@@ -82,7 +82,7 @@ int32_t shiftRightBy(int32_t offset, int32_t a) {
     return a >> offset;
 }
 
-uint32_t shiftRightZfBy(int32_t offset, int32_t a) {
+uint64_t shiftRightZfBy(int64_t offset, int64_t a) {
     /*
      * JS: var _Bitwise_shiftRightZfBy = F2(function(offset, a) { return a >>> offset; });
      *
@@ -94,7 +94,7 @@ uint32_t shiftRightZfBy(int32_t offset, int32_t a) {
      * HELPERS: None
      * LIBRARIES: None (built-in operator)
      */
-    return static_cast<uint32_t>(a) >> offset;
+    return static_cast<uint64_t>(a) >> offset;
 }
 
 } // namespace Elm::Kernel::Bitwise
